Myelin, Basal Ganglia, and Pineal Gland
Investigated the role of myelin as a lipid-based insulator enhancing axonal conduction through saltatory conduction. Studied the significance of nodes of Ranvier in maintaining the integrity of action potentials.

Analyzed the basal ganglia's dual-pathway model in motor control, exploring the balance of direct and indirect circuits in modulating movement via inhibitory and excitatory feedback.

Explored the pineal gland's role in melatonin secretion and circadian rhythm regulation, with insights into its response to environmental light via the suprachiasmatic nucleus and implications for sleep-wake cycles.